Move CompositePixmap to 'pixmap' module
