+++ gridsync/gui/pixmap.py
@@ -2,7 +2,7 @@
 
 
 from PyQt5.QtCore import Qt
-from PyQt5.QtGui import QPixmap
+from PyQt5.QtGui import QColor, QPainter, QPixmap
 
 from gridsync import resource
 
@@ -14,3 +14,22 @@ class Pixmap(QPixmap):
             self.swap(self.scaled(
                 size, size, Qt.KeepAspectRatio, Qt.SmoothTransformation 
             ))
+
+
+class CompositePixmap(QPixmap):
+    def __init__(self, pixmap, overlay=None, grayout=False):
+        super(CompositePixmap, self).__init__()
+        base_pixmap = QPixmap(pixmap)
+        if grayout:
+            painter = QPainter(base_pixmap)
+            painter.setCompositionMode(painter.CompositionMode_SourceIn)
+            painter.fillRect(base_pixmap.rect(), QColor(128, 128, 128, 128))
+            painter.end()
+        if overlay:
+            width = int(base_pixmap.size().width() / 2)
+            height = int(base_pixmap.size().height() / 2)
+            overlay_pixmap = QPixmap(overlay).scaled(width, height)
+            painter = QPainter(base_pixmap)
+            painter.drawPixmap(width, height, overlay_pixmap)
+            painter.end()
+        self.swap(base_pixmap)
