+++ gridsync/gui/widgets.py
@@ -18,25 +18,6 @@ from gridsync.msg import error
 from gridsync.tor import get_tor
 
 
-class CompositePixmap(QPixmap):
-    def __init__(self, pixmap, overlay=None, grayout=False):
-        super(CompositePixmap, self).__init__()
-        base_pixmap = QPixmap(pixmap)
-        if grayout:
-            painter = QPainter(base_pixmap)
-            painter.setCompositionMode(painter.CompositionMode_SourceIn)
-            painter.fillRect(base_pixmap.rect(), QColor(128, 128, 128, 128))
-            painter.end()
-        if overlay:
-            width = int(base_pixmap.size().width() / 2)
-            height = int(base_pixmap.size().height() / 2)
-            overlay_pixmap = QPixmap(overlay).scaled(width, height)
-            painter = QPainter(base_pixmap)
-            painter.drawPixmap(width, height, overlay_pixmap)
-            painter.end()
-        self.swap(base_pixmap)
-
-
 class ConnectionSettings(QWidget):
     def __init__(self):
         super(ConnectionSettings, self).__init__()
